refactor: electron.d.tsをドメイン別に分割し不要なre-exportを削除
- 2,636行のelectron.d.tsを18のドメイン別サブインターフェースに分割
- MyAPIをextendsによる合成インターフェースとして再構成
- 全てのre-exportを廃止し、import元を直接参照に変更（26箇所）
- 消費者ゼロの不要なバレルindex.tsを4件削除
- useScoringSettings.tsのデッドコードre-exportを削除
- useStudentAnswerUpload.ts（消費者ゼロの互換シム）を削除
- 07-score-at-once/types/index.tsの不要なStudentAnswerImageWithDetails re-exportを削除
- CLAUDE.mdにre-export禁止・バレルファイル作成基準のポリシーを追記
