--- src/app/textbox-on-canvas-v4/components/index.ts
@@ -1,10 +1,0 @@
-/**
- * @fileoverview textbox-on-canvas-v4 コンポーネントの統合エクスポート
- * @description 全プレビューコンポーネントの統一インターフェース
- */
-
-export { AnchorControlPanel } from "./AnchorControlPanel"
-export { TextboxCanvasPreview } from "./TextboxCanvasPreview"
-export { TextboxImagePreview } from "./TextboxImagePreview"
-export { TextboxPreview } from "./TextboxPreview"
-export { TextboxSvgPreview } from "./TextboxSvgPreview"
