--- src/app/textbox-on-canvas-v4/hooks/index.ts
@@ -1,7 +1,0 @@
-/**
- * @fileoverview textbox-on-canvas-v4 フックの統合エクスポート
- * @description 全カスタムフックの統一インターフェース
- */
-
-export { useCanvasManagement } from "./useCanvasManagement"
-export { useTextBoxOperations } from "./useTextBoxOperations"
