--- src/components/drawing/utils/index.ts
@@ -1,11 +1,0 @@
-/**
- * @fileoverview 描画ユーティリティのエクスポート
- */
-
-export { type CanvasDrawParams, redrawCanvas } from "./canvasRenderer"
-export {
-  getAbsoluteCoordinates,
-  getRelativeCoordinates,
-} from "./coordinateUtils"
-export { isPointInAnnotation } from "./hitDetection"
-export { drawArrowHead, drawWaveLine, drawZigzagLine } from "./lineRendering"
