+++ src/components/exams/03-region-info/components/RegionDetailsTable.tsx
@@ -7,8 +7,8 @@ import { DeleteConfirmModal } from "@/components/exams/03-region-info/components
 import { RegionTableRow } from "@/components/exams/03-region-info/components/RegionTableRow"
 import { useDragAndDrop } from "@/components/exams/03-region-info/hooks/useDragAndDrop"
 import { useKeyboardNavigation } from "@/components/exams/03-region-info/hooks/useKeyboardNavigation"
-import type { CropRegionWithDetails } from "@/types/electron"
 import type { CropRegionOmrConfigWithOptions } from "@/types/omr.types"
+import type { CropRegionWithDetails } from "@/types/prismaExtensions"
 
 type RegionDetailsTableProps = {
   regions: CropRegionWithDetails[]
