+++ src/components/exams/03-region-info/hooks/useDragAndDrop.ts
@@ -1,6 +1,6 @@
 import { useState } from "react"
 
-import type { CropRegionWithDetails } from "@/types/electron"
+import type { CropRegionWithDetails } from "@/types/prismaExtensions"
 
 type DragState = {
   draggedIndex: number | null
