+++ src/components/exams/05-students/components/exam-students-page/components/ClassStatisticsCards.tsx
@@ -1,6 +1,6 @@
 "use client"
 
-import type { ExamClassWithClass } from "@/types/electron.d"
+import type { ExamClassWithClass } from "@/types/electron/examClassApi"
 
 interface ClassStatisticsCardsProps {
   examClasses: ExamClassWithClass[]
