+++ src/components/exams/07-score-at-once/ScoringMain/ScoreComparisonModal.tsx
@@ -25,10 +25,8 @@ import {
 import { Input } from "@/components/ui/input"
 import { Label } from "@/components/ui/label"
 import { Textarea } from "@/components/ui/textarea"
-import type {
-  QuestionScoreComparisonResult,
-  QuestionScoreWithUser,
-} from "@/types/electron.d"
+import type { QuestionScoreComparisonResult } from "@/types/electron/scoringApi"
+import type { QuestionScoreWithUser } from "@/types/prismaExtensions"
 
 interface ScoreComparisonModalProps {
   isOpen: boolean
