--- src/components/exams/08-export/components/individual-report/index.ts
@@ -1,9 +1,0 @@
-/**
- * 個人成績表プレビューコンポーネント
- * HTML/CSSベースでWYSIWYGプレビューと印刷に対応
- */
-export { BoxPlotChart } from "./BoxPlotChart"
-export { IndividualReportPreview } from "./IndividualReportPreview"
-export { LearningAdvicePreview } from "./LearningAdvicePreview"
-export { generatePrintHtml, injectPrintStyles } from "./printStyles"
-export { ScoreTablePreview } from "./ScoreTablePreview"
