+++ src/components/exams/forms/EditExamWindow.tsx
@@ -16,7 +16,7 @@ import {
 import { Input } from "@/components/ui/input"
 import { Label } from "@/components/ui/label"
 import { Textarea } from "@/components/ui/textarea"
-import type { ExamWithDetails } from "@/types/electron"
+import type { ExamWithDetails } from "@/types/common.types"
 
 type Tag = {
   id: string
