--- src/hooks/useStudentAnswerUpload.ts
@@ -1,5 +1,0 @@
-"use client"
-
-// Re-export types and main hook
-export * from "@/components/exams/06-student-answers/student-answer-management/hooks/types"
-export { useStudentAnswerUploadMain as useStudentAnswerUpload } from "@/components/exams/06-student-answers/student-answer-management/hooks/useStudentAnswerUploadMain"
